Improve error logging of uncaught errors
