+++ src/background/browserAction.ts
@@ -16,7 +16,7 @@
  */
 
 import { isBackground } from "webext-detect-page";
-import { reportError } from "@/telemetry/logging";
+import { reportError } from "@/telemetry/rollbar";
 import { ensureContentScript, showErrorInOptions } from "@/background/util";
 import browser, { Tabs } from "webextension-polyfill";
 import { toggleActionPanel } from "@/contentScript/messenger/api";
@@ -52,7 +52,6 @@ async function handleBrowserAction(tab: Tabs.Tab): Promise<void> {
     updateDevTools({ page: `/devtoolsPanel.html?tabId=${tab.id}` });
   } catch (error) {
     await showErrorInOptions("ERR_BROWSER_ACTION_TOGGLE", tab.index);
-    console.error(error);
     reportError(error);
   }
 }
