+++ src/components/ErrorBoundary.tsx
@@ -20,7 +20,7 @@ import { Button } from "react-bootstrap";
 import { isExtensionContext } from "webext-detect-page";
 import { FontAwesomeIcon } from "@fortawesome/react-fontawesome";
 import { faRedo } from "@fortawesome/free-solid-svg-icons";
-import { reportError } from "@/telemetry/logging";
+import { reportError } from "@/telemetry/rollbar";
 import { getErrorMessage } from "@/errors";
 import { UnknownObject } from "@/types";
