+++ src/devTools/context.ts
@@ -21,7 +21,7 @@ import browser from "webextension-polyfill";
 import { navigationEvent } from "@/background/devtools/external";
 import { useAsyncEffect } from "use-async-effect";
 import { FrameworkMeta } from "@/messaging/constants";
-import { reportError } from "@/telemetry/logging";
+import { reportError } from "@/telemetry/rollbar";
 import { uuidv4 } from "@/types/helpers";
 import { useTabEventListener } from "@/hooks/events";
 import { getErrorMessage } from "@/errors";
