+++ src/permissionsPopup.tsx
@@ -17,8 +17,6 @@
 
 import "@/extensionContext";
 
-// Init rollbar early so we get error reporting on the other initialization
-import "@/telemetry/rollbar";
 import PermissionsPopup from "@/popups/PermissionsPopup";
 
 import ReactDOM from "react-dom";
