+++ src/telemetry/reportUncaughtErrors.ts
@@ -19,7 +19,7 @@
  * @file This file must be imported as early as possible in each entrypoint, once
  */
 
-import { reportError } from "@/telemetry/logging";
+import { reportError } from "@/telemetry/rollbar";
 
 function defaultErrorHandler(
   errorEvent: ErrorEvent | PromiseRejectionEvent
@@ -32,23 +32,15 @@ function defaultErrorHandler(
   }
 
   reportError(errorEvent);
-  errorEvent.preventDefault();
-}
-
-const seen = new WeakSet<ErrorEvent | PromiseRejectionEvent>();
-function avoidLoops(errorEvent: ErrorEvent | PromiseRejectionEvent): void {
-  if (seen.has(errorEvent)) {
-    errorEvent.preventDefault();
-  } else {
-    seen.add(errorEvent);
-  }
 }
 
 /**
  * Set of error event handlers to run before the default one.
  * They can call `event.preventDefault()` to avoid reporting the error.
  */
-export const uncaughtErrorHandlers = new Set([avoidLoops]);
+export const uncaughtErrorHandlers = new Set<
+  (errorEvent: ErrorEvent | PromiseRejectionEvent) => void
+>();
 
 /*
 Refactor beware: Do not add an `init` function or it will run too late.
